+++ src/app/support-request/page.tsx
@@ -1,9 +1,6 @@
+import { SupportRequestStepper } from '@/components/pages/SupportRequest';
+
 export default function SupportRequestPage() {
-  return (
-    <div className="p-4">
-      <h1>Support Request Form</h1>
-      <p>Stepper component will be implemented here.</p>
-    </div>
-  );
+  return <SupportRequestStepper />;
 }