+++ src/components/input/CPTSteps.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+import { Steps, StepsProps } from 'primereact/steps';
+
+export interface CPTStepsProps extends StepsProps {}
+
+export const CPTSteps = (props: CPTStepsProps) => {
+  return <Steps {...props} />;
+};
+