+++ src/components/pages/SupportRequest/index.ts
@@ -1,0 +1,6 @@
+export * from './StepCaseSelection';
+export * from './StepPersonalInfo';
+export * from './StepIssueDetails';
+export * from './StepDescriptionUpload';
+export * from './SupportRequestStepper';
+