+++ src/constants/index.ts
@@ -1,3 +1,2 @@
-// App constants will be exported here
-export {};
+export * from './supportRequest';